// client/src/app/models/cast-response.model.ts
export interface CastDetailsResponse {
  id: number;
  cast: CastInfo[];
  crew: CrewInfo[];
}

export interface CastInfo {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}
[];

export interface CrewInfo {
  adult: boolean;
  credit_id: string;
  department: string;
  gender: number;
  id: number;
  job: string;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
}